requirements.txt
# --- Django 本体と関連ツール ---
Django==5.2.4
# djangorestframework==3.16.0
django-environ==0.12.0
django-axes==8.0.0

# --- API ---
requests

# --- Render用の設定 ---
psycopg2-binary
dj-database-url
whitenoise[brotli]
gunicorn